refactor: deduplicate buildExampleLibrary into shared extension on DartCodeExample
Both analyzer.dart and daco_rule.dart had their own copy of the logic to
compose a DartCodeExample into a ComposedDartBlock. Move the superset
version to block.dart as a public extension so both consumers share it.
